reduced file levels of eav module.
Removed used imports and reduced the file levels of the eav
